fix(rust): satisfy fmt and clippy checks

+++ client/desktop/src-tauri/src/screenshot_guard.rs
@@ -57,7 +57,7 @@ pub fn set_screenshot_protection(enabled: bool) {
 pub fn init_screenshot_protection(window: Window) {
     #[cfg(target_os = "windows")]
     {
-        let hwnd = window.hwnd().unwrap().0 as isize;
+        let hwnd = window.hwnd().unwrap().0;
         let _ = WINDOW_HWND.set(hwnd);
         PROTECTION_ON.store(true, Ordering::SeqCst);
         spawn_guard_thread("sg-affinity-keepalive", move || {
